requirements.txt
pygame>=2.5.0
numpy>=1.24.0
pillow>=10.0.0
psd-tools>=1.9.0
